cmp_complete_diagnostics
