+++ crates/driver/src/db.rs
@@ -4,7 +4,10 @@ use codespan_reporting::term::{
     termcolor::{BufferWriter, ColorChoice},
 };
 use common::file::File;
-use common::{define_input_db, diagnostics::CompleteDiagnostic};
+use common::{
+    define_input_db,
+    diagnostics::{CompleteDiagnostic, cmp_complete_diagnostics},
+};
 use hir::analysis::{
     analysis_pass::{AnalysisPassManager, EventLowerPass, MsgLowerPass, ParsingPass},
     diagnostics::DiagnosticVoucher,
@@ -134,19 +137,7 @@ impl DiagnosticsCollection<'_> {
 }
 
 fn sort_complete_diagnostics(diags: &mut [CompleteDiagnostic]) {
-    diags.sort_by(|lhs, rhs| match lhs.error_code.cmp(&rhs.error_code) {
-        std::cmp::Ordering::Equal => {
-            let lhs_span = lhs.primary_span();
-            let rhs_span = rhs.primary_span();
-            match (lhs_span, rhs_span) {
-                (Some(lhs_span), Some(rhs_span)) => lhs_span.cmp(&rhs_span),
-                (Some(_), None) => std::cmp::Ordering::Less,
-                (None, Some(_)) => std::cmp::Ordering::Greater,
-                (None, None) => std::cmp::Ordering::Equal,
-            }
-        }
-        ord => ord,
-    });
+    diags.sort_by(cmp_complete_diagnostics);
 }
 
 fn sort_and_dedup_complete_diagnostics(diags: &mut Vec<CompleteDiagnostic>) {
