+++ crates/hir/src/analysis/diagnostics.rs
@@ -23,7 +23,7 @@ use crate::{
 };
 use common::diagnostics::{
     CompleteDiagnostic, DiagnosticPass, GlobalErrorCode, LabelStyle, Severity, Span, SpanKind,
-    SubDiagnostic,
+    SubDiagnostic, cmp_complete_diagnostics,
 };
 use either::Either;
 use itertools::Itertools;
@@ -266,19 +266,7 @@ where
     let config = term::Config::default();
 
     let mut completes: Vec<_> = diags.into_iter().map(|diag| diag.to_complete(db)).collect();
-    completes.sort_by(|lhs, rhs| match lhs.error_code.cmp(&rhs.error_code) {
-        std::cmp::Ordering::Equal => {
-            let lhs_span = lhs.primary_span();
-            let rhs_span = rhs.primary_span();
-            match (lhs_span, rhs_span) {
-                (Some(lhs_span), Some(rhs_span)) => lhs_span.cmp(&rhs_span),
-                (Some(_), None) => std::cmp::Ordering::Less,
-                (None, Some(_)) => std::cmp::Ordering::Greater,
-                (None, None) => std::cmp::Ordering::Equal,
-            }
-        }
-        ord => ord,
-    });
+    completes.sort_by(cmp_complete_diagnostics);
 
     for diag in completes {
         term::emit(
