+++ crates/hir/src/test_db.rs
@@ -37,7 +37,7 @@ use codespan_reporting::{
 };
 use common::{
     InputDb, define_input_db,
-    diagnostics::{LabelStyle, Severity, Span},
+    diagnostics::{LabelStyle, Severity, Span, cmp_complete_diagnostics},
     file::File,
     indexmap::IndexMap,
     stdlib::{HasBuiltinCore, HasBuiltinStd},
@@ -254,21 +254,8 @@ impl HirAnalysisTestDb {
             let mut buffer = writer.buffer();
             let config = term::Config::default();
 
-            // copied from driver
             let mut diags: Vec<_> = filtered_diags.iter().map(|d| d.to_complete(self)).collect();
-            diags.sort_by(|lhs, rhs| match lhs.error_code.cmp(&rhs.error_code) {
-                std::cmp::Ordering::Equal => {
-                    let lhs_span = lhs.primary_span();
-                    let rhs_span = rhs.primary_span();
-                    match (lhs_span, rhs_span) {
-                        (Some(lhs_span), Some(rhs_span)) => lhs_span.cmp(&rhs_span),
-                        (Some(_), None) => std::cmp::Ordering::Less,
-                        (None, Some(_)) => std::cmp::Ordering::Greater,
-                        (None, None) => std::cmp::Ordering::Equal,
-                    }
-                }
-                ord => ord,
-            });
+            diags.sort_by(cmp_complete_diagnostics);
 
             for diag in diags {
                 let cs_diag = &diag.to_cs(self);
